requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pyspark==3.5.0
kafka-python==2.0.2
elasticsearch==8.11.0
pandas==2.1.4
numpy==1.26.4
